Update
Work in Progress of further dev. Have sample data set, and store it in an array and print into json format. For the features a constructor has been created and getters and setters. You will notice it has not been implemented into the reader yet. That is what I am currently working on now.