Update: Wrapping Lambda
On 7/13/21 was able to work on application some more. Started to figure out how to get data from bucket save it and write it to lambda /tmp file. Have yet to test this code just uploading for a backup. I have pseudocode/design on paper, and have begun implementing it into what you see here as of right now. Hence to why this has yet to be tested.